move Globe cesium component to a folder that is not located under pulse since it is used in more places
