+++ components/vis/globe-cesium/component.js
