+++ components/vis/globe-cesium/initial-state.js
