+++ components/vis/globe-cesium/reducers.js
