+++ store.js
@@ -37,7 +37,7 @@ import * as exploreDatasetFilters from 'components/app/explore/explore-dataset-f
 import * as layerMenuDropdown from 'pages/app/pulse/layer-menu-dropdown';
 import * as layerCard from 'pages/app/pulse/layer-card';
 import * as layerPill from 'pages/app/pulse/layer-pill';
-import * as globeCesium from 'pages/app/pulse/globe-cesium';
+import * as globeCesium from 'components/vis/globe-cesium';
 import * as pulse from 'pages/app/pulse/pulse';
 
 // Widget
